+++ tests/test_climate.py
@@ -28,6 +28,7 @@ from hahomematic.platforms.custom.climate import (
     ScheduleProfile,
     ScheduleSlotType,
     ScheduleWeekday,
+    _convert_simple_to_weekday,
     _ModeHmIP,
 )
 
@@ -580,4 +581,36 @@ async def test_climate_ip_with_pydevccu(central_unit_mini) -> None:
         profile="P1",
         weekday="MONDAY",
         weekday_data=manual_week_profile_data,
+    )
+
+    manual_simple_weekday_list = [
+        {"TEMPERATURE": 17.0, "STARTTIME": "05:00", "ENDTIME": "06:00"},
+        {"TEMPERATURE": 22.0, "STARTTIME": "19:00", "ENDTIME": "22:00"},
+        {"TEMPERATURE": 17.0, "STARTTIME": "09:00", "ENDTIME": "15:00"},
+    ]
+    weekday_data = _convert_simple_to_weekday(
+        base_temperature=16.0, simple_weekday_list=manual_simple_weekday_list
+    )
+
+    assert weekday_data == {
+        1: {ScheduleSlotType.ENDTIME: "05:00", ScheduleSlotType.TEMPERATURE: 16.0},
+        2: {ScheduleSlotType.ENDTIME: "06:00", ScheduleSlotType.TEMPERATURE: 17.0},
+        3: {ScheduleSlotType.ENDTIME: "09:00", ScheduleSlotType.TEMPERATURE: 16.0},
+        4: {ScheduleSlotType.ENDTIME: "15:00", ScheduleSlotType.TEMPERATURE: 17.0},
+        5: {ScheduleSlotType.ENDTIME: "19:00", ScheduleSlotType.TEMPERATURE: 16.0},
+        6: {ScheduleSlotType.ENDTIME: "22:00", ScheduleSlotType.TEMPERATURE: 22.0},
+        7: {ScheduleSlotType.ENDTIME: "24:00", ScheduleSlotType.TEMPERATURE: 16.0},
+        8: {ScheduleSlotType.ENDTIME: "24:00", ScheduleSlotType.TEMPERATURE: 16.0},
+        9: {ScheduleSlotType.ENDTIME: "24:00", ScheduleSlotType.TEMPERATURE: 16.0},
+        10: {ScheduleSlotType.ENDTIME: "24:00", ScheduleSlotType.TEMPERATURE: 16.0},
+        11: {ScheduleSlotType.ENDTIME: "24:00", ScheduleSlotType.TEMPERATURE: 16.0},
+        12: {ScheduleSlotType.ENDTIME: "24:00", ScheduleSlotType.TEMPERATURE: 16.0},
+        13: {ScheduleSlotType.ENDTIME: "24:00", ScheduleSlotType.TEMPERATURE: 16.0},
+    }
+
+    await climate_bwth.set_profile_weekday_simple(
+        profile="P1",
+        weekday="MONDAY",
+        base_temperature=16.0,
+        simple_weekday_list=manual_simple_weekday_list,
     )